+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderToolbox/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { useEditor, Element } from '@craftjs/core';
 
 // Router
-import { withRouter, WithRouterProps } from 'react-router';
+import { withRouter, WithRouterProps } from 'utils/cl-router/withRouter';
 
 // Intl
 import { injectIntl } from 'utils/cl-intl';
@@ -13,7 +13,6 @@ import { InjectedIntlProps } from 'react-intl';
 // Components
 import ToolboxItem from './ToolboxItem';
 import { Box } from '@citizenlab/cl2-component-library';
-import Container from '../CraftComponents/Container';
 import Text from '../CraftComponents/Text';
 import TwoColumn from '../CraftComponents/TwoColumn';
 import ThreeColumn from '../CraftComponents/ThreeColumn';
@@ -58,21 +57,6 @@ const ContentBuilderToolbox = ({
     >
       <Box w="100%" display="inline">
         <DraggableElement
-          id="e2e-draggable-single-column"
-          ref={(ref) =>
-            ref &&
-            connectors.create(
-              ref,
-              <Element canvas is={Container} id="container" />
-            )
-          }
-        >
-          <ToolboxItem
-            icon="column1"
-            label={formatMessage(messages.oneColumn)}
-          />
-        </DraggableElement>
-        <DraggableElement
           id="e2e-draggable-two-column"
           ref={(ref) =>
             ref &&
@@ -83,7 +67,12 @@ const ContentBuilderToolbox = ({
                 is={TwoColumn}
                 columnLayout="1-1"
                 id="twoColumn"
-              />
+              />,
+              {
+                onCreate: (node) => {
+                  selectNode(node.rootNodeId);
+                },
+              }
             )
           }
         >
@@ -98,7 +87,12 @@ const ContentBuilderToolbox = ({
             ref &&
             connectors.create(
               ref,
-              <Element canvas is={ThreeColumn} id="threeColumn" />
+              <Element canvas is={ThreeColumn} id="threeColumn" />,
+              {
+                onCreate: (node) => {
+                  selectNode(node.rootNodeId);
+                },
+              }
             )
           }
         >
@@ -174,7 +168,12 @@ const ContentBuilderToolbox = ({
             ref &&
             connectors.create(
               ref,
-              <Element is={AboutBox} id="AboutBox" projectId={projectId} />
+              <Element is={AboutBox} id="AboutBox" projectId={projectId} />,
+              {
+                onCreate: (node) => {
+                  selectNode(node.rootNodeId);
+                },
+              }
             )
           }
         >