+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Fixed
 
 - [CL-888] Fixed issue with folders page layout. Project cards are now displayed in two columns (instead of one) on large screens.
+- 
+## 2022-06-13
+
+### Changed
+
+- [TEC-11] Upgraded react-router frontend dependency
+
 
 ## 2022-06-08_2